Reorganize app startup process
